Prevent Array.insert from adding ',' when empty
Previously Array.insert was not able to recognise that a single
pre-existing Whitespace item does not require adding ',' (preceding
comma) when inserting/appending. For example:

```
>>> doc = atoml.parse("a = [  ]")
>>> doc["a"].append("def")
>>> atoml.dumps(doc)  # previous behaviour resulting in invalid syntax:
a = [  , "def"]
```

The changes implemented make sure this does not happen, and that the
syntax is valid when `dump`ing:
```
>>> atoml.dumps(doc)  # a bit messy, but the syntax is valid:
a = [  "def"]
```

This is related to #26 (but I am afraid it is still not entirely solved).