Prevent comments from disappearing when overwriting entries
Previously when overwriting entries of a container, some comments could
disappear. For example:

```
>>> import atoml
>>> example = 'x = "1, 2, 3"'
>>> doc = atoml.loads(example)
>>> x = atoml.array()
>>> x.comment("`x` should be an array")
[]
>>> x.extend(int(i) for i in doc['x'].split(", "))
>>> x.trivia.comment
'# `x` should be an array'
>>> doc['x'] = x
>>> print(doc.as_string())
x = [1, 2, 3]
>>> x.trivia.comment
''
```

Notice that in the example not only the final TOML document does not
contain the comment, but also the array object `x` has its "trivia"
erased after being inserted.

The changes implemented here try to prevent that from happening:
if the item being inserted has a comment, we should preserve it.